move wasm and base implementations to traits; add grpc_client trait
