+++ packages/strategist-utils/src/cosmos/mod.rs
@@ -1,4 +1,5 @@
 pub mod base_client;
 pub mod errors;
+pub mod grpc_client;
 pub mod signing_client;
 pub mod wasm_client;
